time-series: TimeSeriesTable, allow to have more than 2gb of data (#634)
* time-series: add concurrent loading test

* time-series: CompactStringBuffer, use zero as the null marker to avoid the need to write -1 on the whole buffer

For ByteBuffer::allocate, ByteBuffer::allocateDirect, the javadoc guarantees
"each of its elements will be initialized to zero".
For MemoryMappedBuffer, the user controls the file and can create it with zeroes
(if it is extended, virtually all systems extend it with zeros)

* time-series: TimeSeriesTable, allow to have more than 2gib of data

To do this, we use 1 gibi (2^30) *bytes* segment into the table. This allows to
use memory mapped buffer for these segments (it would not have been possible if
we had used buffers of 2^30 *elements*) and to use bitshifts and bitmasks to
compute indices. It's the largest possible value so the overhead is minimized.

Each time series is still constrained on its own to have a number of elements
that fits in an array (~2 billions). We keep the previous fillBuffer
implementations because they allow toArray() to return a single array without
extra copying (and that's up to 16 gb for doubles). If we had removed the
previous fillBuffer implementation, then starting from ~100 million doubles we
would have had to make copies from the multiple 2gib buffers back into one big
array.

* time-series: add Object.requireNonNull(buffer) to CalculatedTimeSeries::fillBuffer

This makes the code more homogeneous (other fillBuffer methods have it).