Security analysis distribution refactoring: dissociate it from security analysis interface.
Currently, distributed security analysis implements the security analysis interface.
It is a little misleading, as some of the inputs and methods of that interface are not correctly
handled by this implementation, for example because they are not properly serializable.

This evolution dissociates explicitly the distributed execution from the security analysis interface,
which is meant to be the SPI for actual computation implementations.

It also refactors the execution handlers, to make the code more clear and testable.

Signed-off-by: Sylvain Leclerc <sylvain.leclerc@rte-france.com>