Remove generics from ContingenciesProviderFactory interface.
The generic return type does not seem to bring anything really useful,
while on the other side it prevents (or makes clumsy) the use of most
fluent APIS which rely on a known return type (streams, optionals ...).

It also causes "unchecked" warnings in implementations.

It seems very unlikely that this change actually breaks some client code.

Signed-off-by: Sylvain Leclerc <sylvain.leclerc@rte-france.com>