Distributed security analysis may now report logs in exception (#807)
On execution error or any exception during the post-processing of distributed security analysis, we now try to collect logs and report them in a ComputationException.
The exception may be inspected in order to retrieve logs. In particular, the security analysis tool will write down logs collected this way when the execution fails.

This evolution makes the 2 calls "run" and "runWithLog" of the distributed security analysis consistent : they will either return a consistent security analysis result, or throw an exception. In both cases, logs may now be retrieved.

Signed-off-by: yichen88 <tang.yichenyves@gmail.com>