time-series: fix CalculatedTimeSeries::split(int newchunksize)
In TimeSeries::split(list<Timeseries> list, int newchunksize), we access only
the first chunkcount elements of the returned list of this method.
Because it was returning chunksize elements instead of chunkcount elements, sometimes
we have more elements than we need and sometimes we have less. When we had more it was not
a problem (size of chunks > number of chunks) and some where just not used, but when we have less,
the array access in TimeSeries::split(list<Timeseries> list, int newchunksize)
would throw IndexOutOfBoundsException

Signed-off-by: Jon Harper <jon.harper87@gmail.com>