loadflow-validation: restore implementation of validators from a5999c533a
This allows to share code between Transformers3WValidation and TransformersValidation.
Other Validators are changed for consistency.

We restore this because otherwise it would be a breaking change.

Note that this is still a breaking change because it deletes the recently added
method with the tableformatterconfig parameter from the interface ValidationWriterFactory.java

Signed-off-by: Jon Harper <jon.harper87@gmail.com>