AbstractToolTest: assert stderr first and don't interpret the empty string as "everything matches"
We assert stderr first because this shows unexpected exceptions to ease debug.
We match the empty string exactly because most tests should not write anything
to stderr. If they do, it's probably an unexpected exception that we want to
display.

Without these two changes, it's very hard to debug these tests.

Signed-off-by: Jon Harper <jon.harper87@gmail.com>