update maven-checkstyle-plugin 3.0.0->3.1.0 (latest) to fix 'mvn compile'
Currently, if you never did 'mvn install', mvn compile fails:
[INFO] IIDM network model implementation .................. FAILURE [  0.016 s]
[ERROR] Failed to execute goal on project powsybl-iidm-impl: Could not resolve dependencies for project com.powsybl:powsybl-iidm-impl:jar:3.1.0-SNAPSHOT: Could not find artifact com.powsybl:powsybl-commons:jar:tests:3.1.0-SNAPSHOT -> [Help 1]

Note: during a release, javadoc-aggregate-jar does the equivalent of 'mvn
compile' and fails. A workaround is to disable checkstyle during release builds
(-P!'checks'). This was not the case before 3.0.0 because the checks profile
was 'activeByDefault=true', meaning it was disabled when -Prelease was used.

This was fixed in the latest version of checkstyle:
[MCHECKSTYLE-353] - Don't resolve any dependencies 6abad1cb750244

Signed-off-by: Jon Harper <jon.harper87@gmail.com>