Create a new factory for testing a network with all equipments
The network is composed of four substations, with 5 voltage levels.
The loadflow converges on the network with Open Loadflow and Hades2.
With Hades2, one of the generators switches from PV to PQ during the computation.
To go further with this test, we could add more transformers.
It was created at first to test the remote measurements are correctly set on all due equipments on the French network.
Thus, it does not contain any 3 windings transformer.

Signed-off-by: Agnès Leroy <agnes.leroy@rte-france.com>