CompletableFutureTask, fix spurious error logging when cancelling
When calling cancel(true), the CompletableFutureTask was throwing
a NullPointerException in its run method because cancel(true)
causes future.get() to throw a CancellationException without a cause (null)
and completeExceptionally requires a non null Exception

the exception was thrown to the executor and the very end. Executors usually catch and
ignore everything (just logs to stdout) so the problem was just ugly logs and a slight
performance drop.

This was a regression from 5bef72fa5ee, the actual code is reverted to the previous version
(except changing throwable to exception to please avoid sonar warnings I guess)
The code change was not logical, we need to unwrap the cause when we have an ExecutionException
and we also need to not unwrap the cause otherwise.

Signed-off-by: HARPER Jon <jon.harper87@gmail.com>